Updated the project to fetch the student information based on ID

+++ src/main/java/com/bhupi/demo/repository/SubjectRepository.java
@@ -3,5 +3,5 @@ package com.bhupi.demo.repository;
 import com.bhupi.demo.entity.Subject;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+public interface SubjectRepository extends JpaRepository<Subject, Integer> {
 }
